pyproject.toml
[tool.poetry]
name = "madr"
version = "0.1.0"
description = ""
authors = ["Henrique Sebastião <contato@henriquesebastiao.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.1"}
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"
psycopg = {extras = ["binary"], version = "^3.2.1"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.6.1"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"
testcontainers = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations', 'data']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'

[tool.taskipy.tasks]
run = 'fastapi dev madr/app.py'
pre_test = 'task lint'
test = 'pytest tests --cov=madr -vv'
post_test = 'coverage html'
cov = 'firefox-developer-edition htmlcov/index.html'
lint = 'ruff check .; ruff check . --diff'
format = 'ruff format .; ruff check . --fix'